server/requirements.txt
Flask==0.11.1
email==4.0.2
inflection==0.2.0
itsdangerous==0.24
mock==1.0.1
mongokit==0.9.1.1
pymongo==2.8.1
redis==2.10.1
requests==2.10.0
pycrypto==2.6.1
short-url==1.2.1
python-slugify==1.2.0